Don't wipe directories within mods, only all the files.

+++ src/main/java/org/spoutcraft/launcher/util/FileUtils.java
@@ -95,6 +95,17 @@ public class FileUtils {
 	 * @throws IOException in case cleaning is unsuccessful
 	 */
 	public static void cleanDirectory(File directory) throws IOException {
+		cleanDirectory(directory, true);
+	}
+
+	/**
+	 * Cleans a directory without deleting it.
+	 *
+	 * @param directory directory to clean
+	 * @param directories true to delete sub directories as well
+	 * @throws IOException in case cleaning is unsuccessful
+	 */
+	public static void cleanDirectory(File directory, boolean directories) throws IOException {
 		if (!directory.exists()) {
 			String message = directory + " does not exist";
 			throw new IllegalArgumentException(message);
@@ -113,6 +124,9 @@ public class FileUtils {
 		IOException exception = null;
 		for (File file : files) {
 			try {
+				if (file.isDirectory() && !directories) {
+					continue;
+				}
 				forceDelete(file);
 			} catch (IOException ioe) {
 				exception = ioe;
